Fix build_covariance crash on Decimal prices from DB
SQLAlchemy returns NUMERIC columns as decimal.Decimal objects, which
pandas stores as object dtype. np.log() cannot operate on Decimal, so
build_covariance crashed immediately after loading prices. Cast the
pivot to float64 before computing log returns.

+++ services/portfolio-builder/app/select.py
@@ -95,7 +95,7 @@ def build_covariance(
     """
     pivot = prices_df.pivot_table(
         index="date", columns="ticker", values="adjusted_close"
-    ).sort_index()
+    ).sort_index().astype(float)  # Decimal from DB → float64 before log returns
 
     if len(pivot) > window_days:
         pivot = pivot.iloc[-window_days:]
